# docs/swagger.yaml
basePath: /api
definitions:
...
  models.DomainConfig:
    properties:
      createdAt:
        type: string
      domainName:
        type: string
      id:
        type: string
      pathPrefix:
        type: string
      projectId:
        type: string
      redirectTo:
        type: string
      sslCertStatus:
        type: string
      updatedAt:
        type: string
    type: object
  models.EnvironmentConfig:
    properties:
      createdAt:
        type: string
      id:
        type: string
      isDefault:
        type: boolean
      name:
        type: string
      projectId:
        type: string
      updatedAt:
        type: string
    type: object
...
  models.Job:
    properties:
      command:
        type: string
      createdAt:
        type: string
      id:
        type: string
      lastOutput:
        type: string
      lastRunAt:
        type: string
      name:
        type: string
      projectId:
        type: string
      schedule:
        type: string
      status:
        type: string
      updatedAt:
        type: string
    type: object
...
  models.TeamEmailSettings:
    properties:
      createdAt:
        type: string
      id:
        type: string
      resendApiKey:
        type: string
      smtpFromAddress:
        type: string
      smtpFromName:
        type: string
      smtpHost:
        type: string
      smtpPassword:
        type: string
      smtpPort:
        type: integer
      smtpUser:
        type: string
      teamId:
        type: string
      updatedAt:
        type: string
      useResend:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Vessl API Documentation
  title: Vessl API
  version: "1.0"
paths:
...
  /api/deployments/{id}/logs:
    get:
      consumes:
      - application/json
      description: GetLogs endpoint
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: GetLogs endpoint
      tags:
      - Deployments
  /api/deployments/{id}/rollback:
    post:
      consumes:
      - application/json
      description: Rollback endpoint
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Rollback endpoint
      tags:
      - Deployments
  /api/domains/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Domain
      parameters:
      - description: Domain ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Delete Domain
      tags:
      - Projects
  /api/environments/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Environment
      parameters:
      - description: Environment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Delete Environment
      tags:
      - Environments
...
  /api/jobs:
    get:
      consumes:
      - application/json
      description: ListProjectJobs endpoint
      produces:
      - application/json
      responses: {}
      summary: ListProjectJobs endpoint
      tags:
      - Jobs
    post:
      consumes:
      - application/json
      description: Create endpoint
      parameters:
      - description: Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Job'
      produces:
      - application/json
      responses: {}
      summary: Create endpoint
      tags:
      - Jobs
  /api/jobs/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Delete Job
      tags:
      - Jobs
    get:
      consumes:
      - application/json
      description: Get Job
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get Job
      tags:
      - Jobs
  /api/jobs/{id}/trigger:
    post:
      consumes:
      - application/json
      description: Run endpoint
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Run endpoint
      tags:
      - Jobs
...
  /api/projects/{id}/deploy:
    post:
      consumes:
      - application/json
      description: DeployProject endpoint
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: DeployProject endpoint
      tags:
      - Projects
  /api/projects/{id}/domains:
    get:
      consumes:
      - application/json
      description: List domains by project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: List domains by project
      tags:
      - Projects
    post:
      consumes:
      - application/json
      description: Create domain
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.DomainConfig'
      produces:
      - application/json
      responses: {}
      summary: Create domain
      tags:
      - Projects
...
  /api/projects/{id}/environments:
    get:
      consumes:
      - application/json
      description: ListByProject endpoint
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: ListByProject endpoint
      tags:
      - Environments
    post:
      consumes:
      - application/json
      description: Create endpoint
      parameters:
      - description: Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.EnvironmentConfig'
      produces:
      - application/json
      responses: {}
      summary: Create endpoint
      tags:
      - Environments
...
  /api/services/{serviceId}/deploy:
    post:
      consumes:
      - application/json
      description: Trigger Deployment
      parameters:
      - description: Service ID
        in: path
        name: serviceId
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Trigger Deployment
      tags:
      - Deployments
  /api/services/{serviceId}/deployments:
    get:
      consumes:
      - application/json
      description: ListServiceDeployments endpoint
      parameters:
      - description: serviceId
        in: path
        name: serviceId
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: ListServiceDeployments endpoint
      tags:
      - AppServices
...
  /api/teams/{teamId}/email_settings:
    get:
      consumes:
      - application/json
      description: Get Team Email Settings
      parameters:
      - description: Team ID
        in: path
        name: teamId
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get Team Email Settings
      tags:
      - Settings
    put:
      consumes:
      - application/json
      description: Save Team Email Settings
      parameters:
      - description: Team ID
        in: path
        name: teamId
        required: true
        type: string
      - description: Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.TeamEmailSettings'
      produces:
      - application/json
      responses: {}
      summary: Save Team Email Settings
      tags:
      - Settings
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"